release: Chromium 2.0.172 (8205209f3b57a69fe811f18771a533bf01407d7f)
Compare: https://github.com/chromium/chromium/compare/f1cc960cda185ea164b809e44cc3e946a996ff36...8205209f3b57a69fe811f18771a533bf01407d7f

+++ chrome/common/ipc_message_unittest.cc
@@ -17,7 +17,7 @@ TEST(IPCMessageTest, Serialize) {
     "#inva://idurl/",
   };
 
-  for (int i = 0; i < arraysize(serialize_cases); i++) {
+  for (size_t i = 0; i < arraysize(serialize_cases); i++) {
     GURL input(serialize_cases[i]);
     IPC::Message msg(1, 2, IPC::Message::PRIORITY_NORMAL);
     IPC::ParamTraits<GURL>::Write(&msg, input);
@@ -47,4 +47,3 @@ TEST(IPCMessageTest, Serialize) {
   void* iter = NULL;
   EXPECT_FALSE(IPC::ParamTraits<GURL>::Read(&msg, &iter, &output));
 }
-
